Added functions to send and receive JSON objects
Added functions to send and receive JSON object from client to server (sendJson for client, and recvJson for server). Functions will need to be implemented into client/server code. sendJson takes in a JSON obj as an argument and serializes it to send to server. recvJson takes in recieved data from the connection stream and reserializes it and returns a JSON obj.